refactor: Data Types, DTOs, Credit Service validator function. Implemented find One Credit by Credit ID and Customer ID.

+++ src/main/kotlin/dio/spring_kotlin_rest_tdd/model/Customer.kt
@@ -3,7 +3,6 @@ package dio.spring_kotlin_rest_tdd.model
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.validation.constraints.Email
-import org.hibernate.validator.constraints.br.CPF
 
 @Entity
 data class Customer(
@@ -11,7 +10,7 @@ data class Customer(
   val id: Long,
   var firstName: String,
   var lastName: String,
-  var income: Int = 100000,
+  var income: Long = 100000,
   val cpf: String,
   var email: String,
   var cep: String,
